[Infra] Cleaned Up Header Files in Route Folder
Continued the header file cleanup effort in the route folder.

Some of these files may need to be revisited in more detail, but got
some of the major header include issues.

Found that some definitions were in the wrong place, so moved them to
the correct implementation file.
