+++ vpr/src/route/check_route.h
@@ -1,8 +1,8 @@
 #ifndef VPR_CHECK_ROUTE_H
 #define VPR_CHECK_ROUTE_H
-#include "physical_types.h"
+
+#include "netlist.h"
 #include "vpr_types.h"
-#include "route_common.h"
 
 void check_route(const Netlist<>& net_list,
                  enum e_route_type route_type,
