+++ vpr/src/route/connection_based_routing.h
@@ -1,10 +1,10 @@
 #pragma once
+
 #include <vector>
 #include <unordered_map>
-#include "route_tree_fwd.h"
-#include "vpr_types.h"
 #include "timing_info.h"
 #include "vpr_net_pins_matrix.h"
+#include "connection_based_routing_fwd.h"
 
 /***************** Connection based rerouting **********************/
 // encompasses both incremental rerouting through route tree pruning
