+++ vpr/src/route/overuse_report.cpp
@@ -1,6 +1,7 @@
 #include "overuse_report.h"
 
 #include <fstream>
+#include "globals.h"
 #include "physical_types_util.h"
 #include "vpr_utils.h"
 #include "vtr_log.h"
