+++ vpr/src/route/parallel_connection_router.cpp
@@ -1,6 +1,7 @@
 #include "parallel_connection_router.h"
 
 #include <algorithm>
+#include "d_ary_heap.h"
 #include "route_tree.h"
 #include "rr_graph_fwd.h"
