+++ vpr/src/route/route_utils.h
@@ -6,7 +6,6 @@
 #include "router_stats.h"
 #include "timing_info.h"
 #include "vpr_net_pins_matrix.h"
-#include "vpr_types.h"
 
 #include "RoutingDelayCalculator.h"
