+++ vpr/src/route/segment_stats.h
@@ -1,1 +1,6 @@
+#pragma once
+
+#include <vector>
+#include "physical_types.h"
+
 void get_segment_usage_stats(std::vector<t_segment_inf>& segment_inf);
